rearranged project to lib to facilitate benchmarking more easily
