+++ src/lib/mod.rs
@@ -1,0 +1,3 @@
+pub mod data;
+pub mod file_ops;
+pub mod parsing;
